requirements.txt
# Core
discord.py>=2.4.0
aiohttp>=3.9.5
websockets>=12.0
pydantic>=2.7
python-dotenv>=1.0.1
loguru>=0.7.2

# Data & crypto
SQLAlchemy>=2.0.30
aiosqlite>=0.20.0
cryptography>=42.0.5

# Optional: testing
pytest>=8.3.0
pytest-asyncio>=0.23.6
